requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
astroid==2.9.3
bcrypt==3.2.0
black==21.12b0
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
flake8==4.0.1
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
isort==5.10.1
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.0.27
passlib==1.7.4
pathspec==0.9.0
pkg_resources==0.0.0
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pyflakes==2.4.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
regex==2021.11.10
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.0
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wrapt==1.13.3
zipp==3.6.0
